refactor: removing netty socket and datagram

+++ framework/tenio-core/src/main/java/com/tenio/core/network/statistic/NetworkReaderStatistic.java
@@ -2,7 +2,7 @@ package com.tenio.core.network.statistic;
 
 public class NetworkReaderStatistic {
 
-	public void updateReadBytes(int numberBytes) {
+	public void updateReadBytes(long numberBytes) {
 		
 	}
